cloudrun/requirements.txt
fastapi
fsspec
gcsfs
google-cloud-logging
google-cloud-storage
zipstream-new